+++ changelog.md
@@ -5,6 +5,7 @@
 ### Added
 
 * `[PowerManagement]` Critial battery level notification to SystemManager.
+* `[Bluetooth]`  Add settings storage to bluetooth related items
 
 ## [0.51.1 2020-12-18]